Refactored pane context menu into its own component
